Skip FailedStorage in background scan
The background job that scans storages must skip failed storages to
avoid potential exceptions, especially when the failed storage comes
from a shared storage where the source is not accessible.

This fix adds not only a check for FailedStorage in the background
scanner but also hardens the FailedStorage class with a new FailedScanner
class to prevent trying to scan into the failed storage.
Note that the latter hardening is not needed directly for the fix
thanks to the extra check in the background job but could be useful to
catch other code paths that might reach the scanner part.